requirements.txt
pandas
requests
tabulate
tldextract
dnspython
urllib3
webdriver-manager
python-whois
beautifulsoup4
selenium